safari-600.7-branch/LayoutTests/fast/shapes/css-shapes-disabled-expected.txt

Test that parsing of css shapes related properties can be disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testInlineStyle("-webkit-shape-margin", "10px") is null
PASS testComputedStyle("-webkit-shape-margin", "10px") is "0px"
PASS testInlineStyle("-webkit-shape-outside", "circle(10px at 10px 10px)") is null
PASS testComputedStyle("-webkit-shape-outside", "circle(10px at 10px 10px)") is "none"
PASS successfullyParsed is true

TEST COMPLETE
